# task4/repo-stat/subscriber/internal/storage/postgres/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Subscription struct {
	ID        int64
	Owner     string
	RepoName  string
	CreatedAt pgtype.Timestamptz
}